requirements.txt
amqp==5.3.1
ansicon==1.89.0
arrow==1.4.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.3
blessed==1.25.0
celery==5.5.3
celery-types==0.23.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.8
django-celery-beat==2.8.1
django-picklefield==3.3
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jinxed==1.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg-pool==3.2.8
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==3.5.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==14.0.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
